Add a function get, get_str, get_int
Implemented a function get():  method that take a key string argument and an optional Callable argument named fn. This callable will be used to convert the data back to the desired format.

Implemented a function get_int(): a method that accepts a string that is the key returns an integer value of the data

Implemented a function get_str(): a method that accepts a string that is the key and returns a string value of the data